rename file places-business and places-controllers for jsdoc
