+++ server/src/routes/admin/index.js
@@ -16,7 +16,7 @@ import {
   importPlaces,
   sendScheduleInspecteurs,
   updatePlaces,
-} from './places.controllers'
+} from './places-controllers'
 import {
   getStatsPlacesExam,
   getStatsResultsExam,
@@ -487,7 +487,7 @@ router.get('/places', verifyRepartiteurDepartement, getPlaces)
  * /admin/places:
  *   post:
  *     tags: ["Administrateur"]
- *     summary: Chargement le planning des inspecteurs
+ *     summary: Chargement du planning des inspecteurs
  *     description: Permet de charger le planning des inspecteurs pour le département actif de l'utilisateur
  *     produces:
  *       - application/json
@@ -495,46 +495,46 @@ router.get('/places', verifyRepartiteurDepartement, getPlaces)
  *       - bearerAuth: []
  *     requestBody:
  *       content:
- *        multipart/form-data:
- *          schema:
- *            type: object
- *            properties:
- *              file:
- *                type: string
- *                format: binary
- *                description: Fichier au format CSV ou XLSX avec la liste des places
- *              departement:
- *                type: string
- *                description: le département actif selectionné par l'utilisateur
- *          example:
- *            file: planning-93.csv
- *            departement: 93
+ *         multipart/form-data:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               file:
+ *                 type: string
+ *                 format: binary
+ *                 description: Fichier au format CSV ou XLSX avec la liste des places
+ *               departement:
+ *                 type: string
+ *                 description: le département actif selectionné par l'utilisateur
+ *           example:
+ *             file: planning-93.csv
+ *             departement: 93
  *     responses:
- *      200:
- *        description: Retour d'une réussite de traitment de fichier
- *        content:
- *          application/json:
- *            schema:
- *              type: object
- *              properties:
- *                fileName:
- *                  type: string
- *                  description: Le nom de fichier en .csv ou .xlsx conternant les places
- *                success:
- *                  type: boolean,
- *                  description: vaut true,
- *                message:
- *                  type: string
- *                  description: Un message compréhensible par l'usager
- *                places:
- *                  type: array
- *                  description: Les messages sur l'état de traitement des places
- *            example:
- *              fileName: planning-93.csv
- *              success: true
- *              message: Le fichier planning-93.csv a été traité pour le departement 93.
- *              places: [
- *                        {
+ *       200:
+ *         description: Retour d'une réussite de traitement de fichier
+ *         content:
+ *           application/json:
+ *             schema:
+ *               type: object
+ *               properties:
+ *                 fileName:
+ *                   type: string
+ *                   description: Le nom de fichier en .csv ou .xlsx contenant les places
+ *                 success:
+ *                   type: boolean,
+ *                   description: vaut true,
+ *                 message:
+ *                   type: string
+ *                   description: Un message compréhensible par l'usager
+ *                 places:
+ *                   type: array
+ *                   description: Les messages sur l'état de traitement des places
+ *             example:
+ *               fileName: planning-93.csv
+ *               success: true
+ *               message: Le fichier planning-93.csv a été traité pour le departement 93.
+ *               places: [
+ *                         {
  *                           "departement": "93",
  *                           "centre": "Bobigny",
  *                           "inspecteur": "5d9b04db0b279f003c677120",
@@ -550,7 +550,7 @@ router.get('/places', verifyRepartiteurDepartement, getPlaces)
  *                           "message": "Place enregistrée en base"
  *                         },
  *                       ]
- *      400:
+ *       400:
  *         description: Fichier manquant
  *         content:
  *           application/json:
@@ -560,18 +560,18 @@ router.get('/places', verifyRepartiteurDepartement, getPlaces)
  *                 - example:
  *                     success: false
  *                     message: Fichier manquant
- *      401:
- *        $ref: '#/components/responses/InvalidTokenResponse'
- *      500:
- *        $ref: '#/components/responses/UnknownErrorResponse'
+ *       401:
+ *         $ref: '#/components/responses/InvalidTokenResponse'
+ *       500:
+ *         $ref: '#/components/responses/UnknownErrorResponse'
  */
 /**
  * Route pour charger le planning des inspecteurs par département
- * Cette route utilise le middleware [verifyRepartiteurDepartement]{@link module:routes/admin/middleware/verify-repartiteur-departement} et le controleur [importPlaces]{@link module:routes/admin/places-controllers}
+ * Cette route utilise le middleware [verifyRepartiteurDepartement]{@link module:routes/admin/middleware/verify-repartiteur-departement} et le controleur [importPlaces]{@link import('./places-controllers')..importPlaces}
  * @name Router POST '/admin/places'
  *
- * @see [importPlaces]{@link module:routes/admin/places-controllers}
- * @see {@link http://localhost:8000/api-docs/#/Administrateur/post_admin_places}
+ * @see {@link import('./places-controllers')..importPlaces|Fonction importPlaces}
+ * @see {@link http://localhost:8000/api-docs/#/Administrateur/post_admin_places|Swagger: POST /admin/places}
  */
 router.post('/places', verifyRepartiteurDepartement, importPlaces)
 router.delete('/places', deletePlacesByAdmin)
