+++ server/src/routes/admin/places-controllers.spec.js
@@ -42,7 +42,7 @@ import {
   deletePlacesByAdmin,
   getPlaces,
   updatePlaces,
-} from '../admin/places.controllers'
+} from './places-controllers'
 import { SUBJECT_CONVOCATION } from '../business'
 import { getConvocationBody } from '../business/build-mail-convocation'
 import { REASON_MODIFY_RESA_ADMIN } from '../common/reason.constants'
